Lucas/fix prem days and outfit (#16)
* Fix outfit mapping to proto

Outfits were wrongly mapped to proto, so its colors were misplaced. This will properly map the proto properties to the correct database columns.

* Fix uint overflow in premium_unti parameter

PremUntil contains a timestamp of the day when the premium is over. That's a number bigger than uint_32 limit, thus it was overflowing and returning wrong value. This commit updates the protobuf, the structures and methods to uint64.

+++ src/api/models/session.go
@@ -9,7 +9,7 @@ type Session struct {
 	IsReturner                    bool   `json:"isreturner"`
 	LastLoginTime                 uint32 `json:"lastlogintime" proto:"LastLogin"`
 	OptionTracking                bool   `json:"optiontracking"`
-	PremiumUntil                  uint32 `json:"premiumuntil"`
+	PremiumUntil                  uint64 `json:"premiumuntil"`
 	ReturnerNotification          bool   `json:"returnernotification"`
 	SessionKey                    string `json:"sessionkey"`
 	ShowRewardNews                bool   `json:"showrewardnews"`
